*: Enable flush splits in L0
This change updates L0 flush logic to split SSTables when writing
to L0 along flushSplitKeys as calculated by L0SubLevels. This
should lead to fewer wide SSTables in practice. The required
relaxation of L0 invariants (to allow for SSTs that overlap in
sequence numbers) has happened in past changes. Range tombstones
are truncated at these split keys, using FlushToNoStraddling.

A new option, FlushSplitBytes, sets the target number of bytes
to aim for in each flush split interval. Defaults to 0, or flush
splitting disabled.

+++ internal/manifest/l0_sublevels.go
@@ -233,7 +233,7 @@ func insertIntoSubLevel(files []*FileMetadata, f *FileMetadata) []*FileMetadata 
 // IsIntraL0Compacting. Those fields are accessed in InitCompactingFileInfo
 // instead.
 func NewL0SubLevels(
-	files []*FileMetadata, cmp Compare, formatKey base.FormatKey, flushSplitMaxBytes uint64,
+	files []*FileMetadata, cmp Compare, formatKey base.FormatKey, flushSplitMaxBytes int64,
 ) (*L0SubLevels, error) {
 	s := &L0SubLevels{cmp: cmp, formatKey: formatKey}
 	s.filesByAge = files
@@ -313,7 +313,7 @@ func NewL0SubLevels(
 	var cumulativeBytes uint64
 	for i := 0; i < len(s.orderedIntervals); i++ {
 		interval := &s.orderedIntervals[i]
-		if flushSplitMaxBytes > 0 && cumulativeBytes > flushSplitMaxBytes &&
+		if flushSplitMaxBytes > 0 && cumulativeBytes > uint64(flushSplitMaxBytes) &&
 			(len(s.flushSplitUserKeys) == 0 ||
 			!bytes.Equal(interval.startKey.key, s.flushSplitUserKeys[len(s.flushSplitUserKeys)-1])) {
 			s.flushSplitUserKeys = append(s.flushSplitUserKeys, interval.startKey.key)
